+++ threedigrid_builder/constants.py
@@ -3,6 +3,10 @@ from enum import IntEnum
 from enum import unique
 
 
+# The WKT spec to use in the output crs_wkt field:
+WKT_VERSION = "WKT2_2015"
+
+
 @unique
 class CalculationType(IntEnum):
     # channels have +100 in the calculation type, this is mapped on read