+++ threedigrid_builder/interface/__init__.py
@@ -12,6 +12,5 @@ The interface layer depends on the domain layer only.
 from .db import *  # NOQA
 from .geopackage import *  # NOQA
 from .gridadmin import *  # NOQA
-from .raster_rasterio import *  # NOQA
 from .raster_gdal import *  # NOQA
 from .dict_out import *  # NOQA