+++ threedigrid_builder/interface/gridadmin.py
@@ -76,6 +76,10 @@ def field_to_h5(group, name, dtype, val, mode="attrs"):
     None values are converted automatically to -9999 (int an IntEnum), NaN (float)
     and "" (str). bool values cannot be None. Unknown datatypes are skipped silently.
     """
+    # backwards compat: epsg code is saved as string
+    if name == "epsg_code":
+        dtype = str
+        val = str(val)
     # convert Optional[<type>] to <type>
     dtype = unpack_optional_type(dtype)
     # handle Tuple