+++ threedigrid_builder/tests/test_gridadmin.py
@@ -259,7 +259,8 @@ def test_write_meta(h5_out, dataset, shape, dtype):
 @pytest.mark.parametrize(
     "attr,shape,dtype",
     [
-        ("epsg_code", (), "S"),  # changed to int
+        ("epsg_code", (), "S"),
+        ("crs_wkt", (), "S"),  # added
         ("has_1d", (), "bool"),  # changed to bool
         ("has_2d", (), "bool"),  # changed to bool
         ("has_0d", (), "bool"),